Add backend insight story card
