+++ client/web/src/SourcegraphWebApp.scss
@@ -95,6 +95,7 @@ body,
 @import './site-admin/init/SiteInitPage';
 @import './site-admin/SiteAdmin';
 @import './tree/Tree';
+@import './insights/insights-global-styles';
 @import './nav/GlobalNavbar';
 @import './nav/NavLinks';
 @import './auth/ResetPasswordPage';
