+++ client/web/src/insights/components/insights-view-grid/components/extension-insight/ExtensionInsight.tsx
@@ -11,7 +11,7 @@ import { Settings } from '../../../../../schema/settings.schema'
 import { InsightsApiContext } from '../../../../core/backend/api-provider'
 import { ViewInsightProviderSourceType } from '../../../../core/backend/types'
 import { useDeleteInsight } from '../../../../hooks/use-delete-insight/use-delete-insight'
-import { useParallelRequests } from '../../../../hooks/use-parallel-request'
+import { useParallelRequests } from '../../../../hooks/use-parallel-requests/use-parallel-request'
 import { InsightViewContent } from '../../../insight-view-content/InsightViewContent'
 import { InsightErrorContent } from '../insight-card/components/insight-error-content/InsightErrorContent'
 import { InsightLoadingContent } from '../insight-card/components/insight-loading-content/InsightLoadingContent'
