+++ client/web/src/insights/insights-global-styles.scss
@@ -1,0 +1,5 @@
+/**
+    All code insight related non-css-module styles that must be included into the main css bundle
+    via scss @import in SourcegraphWebApp.scss file.
+ */
+@import './components/insight-view-content/InsightViewContent';
